cli: refactor security-status to the new approach
