+++ features/security_status.feature
@@ -79,8 +79,7 @@ Feature: Security status command behavior
     When I verify that running `pro security-status --format unsupported` `as non-root` exits `2`
     Then I will see the following on stderr:
       """
-      usage: security-status [-h] [--format {json,yaml,text}]
-                             [--thirdparty | --unavailable | --esm-infra | --esm-apps]
+      usage: pro security-status [flags]
       argument --format: invalid choice: 'unsupported' (choose from 'json', 'yaml', 'text')
       """
 
@@ -380,8 +379,7 @@ Feature: Security status command behavior
     When I verify that running `pro security-status --thirdparty --unavailable` `as non-root` exits `2`
     Then I will see the following on stderr
       """
-      usage: security-status [-h] [--format {json,yaml,text}]
-                             [--thirdparty | --unavailable | --esm-infra | --esm-apps]
+      usage: pro security-status [flags]
       argument --unavailable: not allowed with argument --thirdparty
       """
     When I run `rm /var/lib/apt/periodic/update-success-stamp` with sudo
@@ -697,8 +695,7 @@ Feature: Security status command behavior
     When I verify that running `pro security-status --thirdparty --unavailable` `as non-root` exits `2`
     Then I will see the following on stderr
       """
-      usage: security-status [-h] [--format {json,yaml,text}]
-                             [--thirdparty | --unavailable | --esm-infra | --esm-apps]
+      usage: pro security-status [flags]
       argument --unavailable: not allowed with argument --thirdparty
       """
     When I run `rm /var/lib/apt/periodic/update-success-stamp` with sudo
